Update ServerInterface
added the push and pop methods for each the send and recieve vectors. Beginning to add switch statements. Need to implement a Messages.java file for constant message code names rather than using direct numbers. This way, it is much easier to read and write to when others begin implementing their packages.