voce added + bug fix
The following commit adds all voce functionality to the voce package. All paths are relative to enable cross environment use.

In addition testing has shown that line 307 of the previous iteration attempted to parse the table id into an integer. This may have show successful in testing but only when there was a table to show. After restarting the server clear of tables. An attempt to start would produce a crash reporting a parse error as the client attempted to convert ""(the empty string) into an integer. In order to remedy this, I added an "if" statement to line 306 in Baac.java stating the table will not be parsed and added if it is equal to the empty string.