Quick n Dirty Jar Creation is almost finished
Easy Jar Creation is almost Finished. After messing around with “JARS within JARS” for about 4 hours. I finally realized the feat was not easily done. After that I realized I would put the classes manually in the src from the jar. I tried that. It failed miserably. The issue is that external resources such as text files get a little fucked up in the creation of a jar. As in they cease to exist to be average files that can be referenced by normal means. So this being the case voce was unable to find or distinguish the vital config files. After that I attempted to manually write the config files within the program. Well that worked briefly until i realized that the xml was pointing to was voce classes needed to be executed. So my final solution was to host the “src.zip” on my private server.  Most of the code for this is written but toggled out at the moment. What happens is that BAAC, prior to starting, downloads the zip file, unzips it on the desktop. The zip is approximately 18.5MB, this means that the total number of bits is 148 Megabits. With as outgoing speed of 5 Megabits per second out of my house, the download of the file will optimally by 29.6 seconds. If this speed is insufficient we could rent out temporary server space on a higher end internet connection. Like I said, more testing has to be done but thats what tomorrow is for. haha.